:recycle: refactor: 회원 삭제 수정 (2023.07.22 김현희)

+++ server/src/main/java/mainproject/cafeIn/domain/member/service/MemberService.java
@@ -199,7 +199,8 @@ public class MemberService {
 
         Member findMember = findById(id);
         if (passwordEncoder.matches(password, findMember.getPassword())) {
-            imageService.delete("profiles", findMember.getImage());
+
+            if (findMember.getImage() != null) imageService.delete("profiles", findMember.getImage());
             memberRepository.deleteFollowerOrFollowing(findMember);
             deletePostBookmarks(findMember.getId());
             memberRepository.deleteCafeBookMarkList(findMember);
